Allow manually configuring the hostname
Without any obvious ability for the extension to determine the machine's
hostname automatically, the best available option seems to be to allow
users to configure it manually. A field is added to the extension popup
to configure the hostname.

The field is blank by default, and the updates were designed to ensure
that the behaviour if the hostname is not set will be unchanged (so
people using workflows that expect browser data in the "unknown" host
bucket won't see anything break; they can set the hostname once they're
prepared to do so).

Fixes #132